refactor(app): extract plot bookmark actions
